Added store module and module validation

+++ src/employee/employee.controller.ts
@@ -6,11 +6,16 @@ import {
   UpdateEmployeeDto,
   ListQuery,
   EmployeeTopics,
+  badRequest,
 } from 'store-mag-types';
+import { StoreService } from 'src/store/store.service';
 
 @Controller('employee')
 export class EmployeeController {
-  constructor(private readonly service: EmployeeService) {}
+  constructor(
+    private readonly service: EmployeeService,
+    private readonly storeService: StoreService,
+  ) {}
 
   @MessagePattern(EmployeeTopics.LIST_EMPLOYEE)
   async list(@Payload() query: ListQuery) {
@@ -19,21 +24,47 @@ export class EmployeeController {
 
   @MessagePattern(EmployeeTopics.CREATE_EMPLOYEE)
   async create(@Payload() data: CreateEmployeeDto) {
+    if (data.storeId) {
+      const store = await this.storeService.findById(data.storeId);
+
+      if (!store) {
+        return badRequest('Store is not correct');
+      }
+    }
+
     return await this.service.create(data);
   }
 
   @MessagePattern(EmployeeTopics.UPDATE_EMPLOYEE)
   async update(@Payload() data: UpdateEmployeeDto) {
+    const employee = await this.service.get(data.id);
+
+    if (!employee) {
+      return badRequest('Employee not found');
+    }
+
     return await this.service.update(data);
   }
 
   @MessagePattern(EmployeeTopics.DELETE_EMPLOYEE)
   async delete(@Payload() id: string) {
+    const employee = await this.service.get(id);
+
+    if (!employee) {
+      return badRequest('Employee not found');
+    }
+
     return await this.service.delete(id);
   }
 
   @MessagePattern(EmployeeTopics.FIND_EMPLOYEE)
   async findById(@Payload() id: string) {
-    return await this.service.get(id);
+    const employee = await this.service.get(id);
+
+    if (!employee) {
+      return badRequest('Employee not found');
+    }
+
+    return employee;
   }
 }
